Title: 정수 두 개의 합 2, Time: 90ms, Memory: 9MB, Status: Wrong Answer - Codetree

+++ 251026/정수 두 개의 합 2/sum-of-two-integers-2.java
@@ -45,7 +45,7 @@ class Solver {
         int i = 0;
         int j = 0;
         int answer = 0;
-        int curSum = numbers.get(i) + numbers.get(j);
+        long curSum = numbers.get(i) + numbers.get(j);
         while (true) {
             while (true) {
                 if (numbers.size() == j + 1) {
